requirements.txt
pandas
matplotlib
Pillow
typing
# tensorflow==1.13
